04-How-to-embed-Datalens-into-your-website-using-secure-embedding-technology/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.114.1
h11==0.14.0
idna==3.8
pydantic==2.9.1
pydantic_core==2.23.3
python-json-logger==2.0.7
setuptools==74.1.2
sniffio==1.3.1
typing_extensions==4.12.2
uvicorn==0.30.6
wheel==0.38.4
requests==2.32.3
mangum~=0.17.0
cryptography~=43.0.3
pyjwt~=2.9.0